bcm27xx: (APP-1062) make FS large enough to destroy overlay
Otherwise, on upgrade we retain the overlay if the rootfs size is the
same (causing confusing behaviour). This was taken from
https://github.com/openwrt/openwrt/issues/9113#issuecomment-1572846457

However, ensure we only clear the bit at the end of the partition to
ensure we don't generate unnecessarily large images for the EKH01.
See https://github.com/openwrt/openwrt/pull/13793

Approved-by: Milad Mohtashamirad
Approved-by: Arien Judge
Approved-by: Evan Benn