keep.d: (APP-1962) Remove config.txt and cmdline from keep.d
All filenames in keep.d are retained if the user chooses 'retain config'
in an update. There is no way for the update to override this at update
time. We may want to modify the command line or bootloader config in
a future update, and as we are supporting 'retain configs' we must
remove these files from keep.d. This will only affect updates FROM the
image that includes this patch.

Approved-by: Arien Judge