os-release: (PR #181) Fill BUILD_ID from git HEAD instead of fixed string
This is a partial revert of
863f769b75beb17b1965786f473b700e0ba0f879

Remove version and version.date which are the fixed string versions from
the 22.03.5 tag. This causes script/getversion.sh to read the revision
from the git log, and similar for the date.

This only modifies the BUILD_ID in os-release, due to other change in
863f769b75beb17b1965786f473b700e0ba0f879

Tested by building an image for EKH01-03.

Approved-by: James Haggerty
Approved-by: Arien Judge