hostapd: (APP-2687, PR #210) Make beacon interval an iface config
Beacon interval can be configured per interface in case of having multiple vifs
configured on a single radio. The generic hostapd allows this to be a per BSS
config. In that way, all BSS either AP/IBSS/Mesh could configure its beacon
interval.

This design was modified in openwrt and it was made to be a common wifi-device
config in a 13 year old commit

https://openwrt-devel.openwrt.narkive.com/K7I3osR6/patch-mac80211-allow-ap-configuration-of-beacon-interval-dtim-period-maximum-permissible-sta-listen-

Approved-by: James Haggerty
Approved-by: Evan Benn