base-files: (APP-2846, PR #226) Replace openwrt banner with Morse one
Previously, this was done in the theme. With the new Argon theme,
we didn't have this, and it was ugly to either copy the banner
over the top (since base-files 'owns' the banner) or mutate
the existing banner.

Doing it this way means:
- we see this banner regardless of the installed theme
- we include the version info
- there's no confusing banner over-writing/mutating

But has the disadvantage that we carry an extra commit
that we'll never upstream.

e.g.

-----------------------------------------------------
 OpenWrt 23.05.2, Morse-2.6-dev
-----------------------------------------------------


Approved-by: Arien Judge
Approved-by: Milad Mohtashamirad