procd: (APP-3185, PR #268) allow button timeouts to timeout
This allows visual feedback when timeouts are reached.

In our case, there is a single button which on a normal
press/release will initiate DPP, but longer presses will
first reboot then factory reset. So users know they've
held it down long enough, it's nice to change a status LED.

This also fixes what seems to me a bug, in that you can
attach a timeout to a 'release' action but the timing was
reset (so you can't distinguish between a released timeout
vs a pressed timeout).

See:

https://github.com/openwrt/procd/pull/10


Approved-by: Sophronia Koilpillai
Approved-by: Arien Judge