package: netifd: (APP-3507, PR #307) aggressively renew DHCP leases on bridge
This causes DHCP renewals when you disconnect/reconnect cables or
re-establish wifi connections on a bridge.

https://github.com/openwrt/netifd/pull/35

I think we want this, but it's a bit of a scary change that I
might be missing some implications for, so I'd like to get it
in now to give us some time to experience it.

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn