bcm27xx: (APP-3631, PR #317) custom distroconfig for ekh devices only
Previously, we'd just been modifying the distroconfig.txt for
all RPIs. This has the advantage that we'd pick up any upstream
changes.

However, for SDK users this is potentially annoying, as they
may want their own set of customisations. It's also a bit
strange that a 'normal' RPI build would have our HAT overlay.

This changes it so we have a custom distroconfig.txt for
ekh01 devices only which completely replaces the base
distroconfig.txt.

Approved-by: Arien Judge
Approved-by: Evan Benn