linux: (APP-3627, PR #315) add invert ability to LED netdev trigger
When you have an RGB LED with every colour mapped individually,
it's useful to have a different colour for link on vs link off,
particularly in the case where the LED has multiple meanings
(i.e. it's a generic status/power LED as well, so shouldn't just be
turned off in the absence of a link).

This invert code is largely copied from similar code
in other triggers which already support invert.

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn