mt76: (APP-1704, APP-3841, PR #346) set insertion order for mt76 wifi module
The reason for this change is to ensure the wifi module is inserted before morse.
This reintroduces a change which we 'lost' in the rebase as theoretically
not necessary, but found it introduced a regression.

Putting the module this early in the load ensures that phy0 is reliably allocated
to the mt76 module; we were finding that it was variable due to mt7603/morse being next to each other alphabetically, which meant that the LED (set to phy0-ap0)
would sometimes not come up.

Ideally, we would have either a generic way of determining module order for
a particular board, OR make the LED code smarter such that you could
specify a device path or similar instead of an ifname, but both of these
are large changes that do not belong in a point release.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn