boards: (APP-4028 PR #353): Add ekh-ath79 board for gl-x750 and glinet gl-ar300m16
Also known as glinet spitz. The mips processor is big endian, so the
current release driver 1.14.1 will not work. I have however tested this
with the big endian patchset and it works.

The board has only 16MB flash, so I have restricted the packages.

The usb port is disabled by a gpio in the upstream dts. I have edited
that to enable the port at boot.

Approved-by: James Haggerty
Approved-by: Lyall Beveridge