boards: (PR #405) add kexec support to dev builds
This lets you load a kernel from userspace.

Why is this useful? Because you might want to pass modparams
to a kernel, and OpenWrt's kernel ignores modparams passed
via uboot (presumably since they often don't control the uboot).

Note that this is not necessary for an ekh01/rpi, since you
can set the modparams via the kernel command line in
/boot/config.txt


Approved-by: Andrew Dodd
Approved-by: Lyall Beveridge