APP-2787 PR #421: Increase ubus message size for iwinfo assoclist
When 2000 stations are connected the iwinfo assoclist command returns
a large amount of data. Use the head for the buffer in the rpcd iwinfo plugin,
the buffer is increased in a separate iwinfo patch.

Also increase the ubus max message size. 2000 stations sends a ~1.7MB
json message. This patch does not work for 8000 stations, which will
likely require a more well thought out fix.

Approved-by: James Haggerty
Approved-by: Lyall Beveridge