morse_setup.sh: (APP-4688, PR #438) make ALL_KMODS optional
There are two downsides to all kmods:
- they increase build times when you're switching kernels
- you sometimes find odd dependency issues in upstream
  which you don't want to fix

This change makes the minimal (-m) build config not use
ENABLE_ALL_KMODS. i.e. the defaults are the same, but using
'-m' the build time is significantly decreased.

We also amend the config application code so that
it's possible to override (almost) anything in
target_diffconfig.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn