base-files: (APP-5613, PR #499) add a new mechanism so we can run defaults code early in boot
We want to setup before kmodloader runs so that when
cfg80211 (regulatory.db) and mt7603e (txpat) load then
load the correct files from /lib/firmware.

If we left this to after kmodloader, we would have to
force these modules to reload, and a new phy number
would be allocated for mt7603e.


Approved-by: Sophronia Koilpillai
Approved-by: Prabha Murugesan