boards/ekh-mt76x8: (APP-5773, PR #515) remove camera app
We are running very close to the maximum disk space on these devices,
and this can cause a poor user experience (accidentally filling
the disk with a minimal number of writes). It also prevents
users for installing ipks.

Not to mention that the camera app can't be heavily used because
it will put too much pressure on the CPU/RAM.

Approved-by: Sophronia Koilpillai
Approved-by: Andrew Dodd
Approved-by: Lyall Beveridge