boards: (PR #529) make sure halowlink devices have halowlink stuff
The problem is that without putting the 'key'
packages in target_diffconfig, a minimal build (i.e. -m) will
leave them out. It's a bit surprising to bring up a halowlink
without any wizards or persistent vars, and this is not consistent
with our EKH devices.

So, despite the duplication, for now I think it's worth moving
this configuration into the target_diffconfig.


Approved-by: Sophronia Koilpillai
Approved-by: Andrew Dodd
Approved-by: Lyall Beveridge