morse_setup.sh: (APP-5942, PR #544) remove real release number from untagged builds
Instead of building 2.12-dev or similar, we'll now build
commithash-date. This is to avoid any confusion with 'real'
releases.

This is done in setup.sh so we avoid hacking OpenWrt Makefiles
or having to modify our release script logic, even though it's a
little weird for it to be here.

Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge